Renamed ConversionContext to RenderContext
