+++ test/ArgvParser.ts
@@ -13,7 +13,14 @@ export class ArgvParser {
      * @param argv   Arguments from process.argv.
      */
     public constructor(argv: string[]) {
-        this.argv = argv;
+        for (const arg of argv) {
+            if (arg.indexOf("--argv=") === 0) {
+                this.argv = JSON.parse(arg.slice("--argv=".length));
+                return;
+            }
+        }
+
+        this.argv = [];
     }
 
     /**
