+++ test/unit/Commands/BannedKeywordsBagTests.ts
@@ -1,10 +1,9 @@
 import { expect } from "chai";
 import "mocha";
 
-import { BannedKeywordsBag } from "../../../lib/Commands/BannedKeywordsBag";
-import { LiteralCommand } from "../../../lib/Commands/LiteralCommand";
-import { ConversionContext } from "../../../lib/Conversions/ConversionContext";
-import { TypeScript } from "../../../lib/Languages/TypeScript";
+import { BannedKeywordsBag } from "../../../lib/Rendering/Commands/BannedKeywordsBag";
+import { LiteralCommand } from "../../../lib/Rendering/Commands/LiteralCommand";
+import { TypeScript } from "../../../lib/Rendering/Languages/TypeScript";
 
 describe("BannedKeywordsBag", () => {
     describe("renderBannedKeyword", () => {
