+++ test/unit/Conversions/Casing/PythonImportCaseConverterTests.ts
@@ -1,17 +1,17 @@
 import { expect } from "chai";
 import "mocha";
 
-import { PythonImportCaseConverter } from "../../../../lib/Conversions/Casing/PythonImportCaseConverter";
+import { PythonImportCaseConverter } from "../../../../lib/Rendering/Casing/PythonImportCaseConverter";
 import { itConvertsFromTo } from "./ConverterTests";
 
 describe("PythonCaseConverter", () => {
     describe("convert", () => {
-        // internal directories
+        // Internal directories
         itConvertsFromTo(PythonImportCaseConverter, ["abc"], "abc");
         itConvertsFromTo(PythonImportCaseConverter, ["abc", "def"], "abc.def");
         itConvertsFromTo(PythonImportCaseConverter, ["abc", "def", "ghi"], "abc.def.ghi");
 
-        // parent directories
+        // Parent directories
         itConvertsFromTo(PythonImportCaseConverter, ["..", "ghi"], ".ghi");
         itConvertsFromTo(PythonImportCaseConverter, ["..", "..", "ghi"], "..ghi");
     });
